fix(object): split keys lookup to a sibling, share the payload offset, canonical handle-band predicate
Three gates: object/mod.rs went 74 lines over the 2000 cap; the new
`(key as usize) < 0x10000` was a bare handle-floor site the addr-class
ratchet rejects; and open-coding the StringHeader payload offset a second
time raised the payload-access ratchet. The two key helpers move together
so the shared offset helper sits with them.
